feat(models): assign FREE plan on tenant creation
Fixes #172
