+++ accounts/signals.py
@@ -1,2 +1,6 @@
-# from django.db.models.signals import post_save
-#  from django.dispatch import receiver
+# from django.db import transaction
+# from django.db.models.signals import post_save, pre_save
+# from django.dispatch import receiver
+#
+# from accounts.models import Tenant, TenantQuota
+
